Products, Index, Contact Us
Added Muhanned's changes to the Default and added the Contact Us Page,
and images required for his pages.

Changed all buttons on product page to labels.  Added a list for product
information to be pulled for the cart, need to update the buttons to
insert the information to the labels.