Design Changes
Changed the color of the pngs to blueish
Added a container to the nav bar.
Made images on home page responsive
Changed back button to use bootstrap
Added a bootstrap carousel for pictures on home page
changed  some more images to image buttons. Took me a while to figure
out the naming conventions in the back end.